# images-api/src/tests/post-deployment/downloaded-app-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Images API",
    "description": "Images API - https://github.com/connected-web/gfx-suite",
    "version": "2024-07-15T11:20:26Z"
  },
  "servers": [
    {
      "url": "https://images.dev.connected-web.services"
    }
  ],
  "paths": {
    "/user/{userId}": {
      "get": {
        "operationId": "userByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Credentials": {
                "schema": {
                  "type": "string"
                }
              },
              "Content-Type": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetailsModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "ImagesAPIImagesAPIPrivateApiRequestAuthorizerEC37AB56": []
          }
        ]
      },
      "options": {
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "204 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Credentials": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserDetailsModel": {
        "title": "User Details Response",
        "required": [
          "message",
          "user"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "The message content describing the response"
          },
          "user": {
            "required": [
              "decryptionKey",
              "userId"
            ],
            "type": "object",
            "properties": {
              "decryptionKey": {
                "type": "string",
                "description": "The user's decryption key for secure data storage"
              },
              "userId": {
                "type": "string",
                "description": "The unique identifier for the user"
              }
            },
            "description": "The user details"
          }
        }
      },
...
    },
    "securitySchemes": {
      "ImagesAPIImagesAPIPrivateApiRequestAuthorizerEC37AB56": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "x-amazon-apigateway-authtype": "custom"
      }
    }
  }
}